đã fix 1 vài bug đã liệt kê (tào lao chapter 1) - còn 2 bug đang cân nhắc

+++ MyWebPlay/Controllers/AdminController2.cs
@@ -1249,6 +1249,7 @@ namespace MyWebPlay.Controllers
                     // Lấy danh sách tất cả các file và thư mục con
                     foreach (FileInfo file in dir.GetFiles())
                     {
+                        if (file.Exists)
                         file.Delete(); // Xoá từng file
                     }
